pyproject.toml
[project]
name = "browser-pilot"
version = "0.1.0"
description = "Autonomous browser agent using Ollama vision, Playwright, and LangChain"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "Chirag Singhal", email = "whyiswhen@gmail.com" }]
dependencies = [
    "browser-use>=0.12.5,<1.0",
    "playwright>=1.58.0,<2.0",
    "langchain>=0.3.25,<1.0",
    "langchain-core>=1.2.23,<2.0",
    "langchain-openai>=0.3.12,<1.0",
    "fastapi>=0.135.2,<1.0",
    "uvicorn[standard]>=0.34.0,<1.0",
    "pydantic>=2.10,<3.0",
    "pydantic-settings>=2.13.1,<3.0",
    "pillow>=11.1,<12.0",
    "aiohttp>=3.11,<4.0",
    "structlog>=24.4,<25.0",
    "python-dotenv>=1.0,<2.0",
    "click>=8.1,<9.0",
    "rich>=13.0,<14.0",
    "sse-starlette>=2.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<1.0",
    "pytest-playwright>=0.6,<1.0",
    "ruff>=0.9,<1.0",
    "coverage>=7.6,<8.0",
]

[project.scripts]
browserpilot = "browser_pilot.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/browser_pilot"]

[tool.ruff]
line-length = 80
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["browser_pilot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]
